requirements.txt
Pillow>=10.0
pillow-heif>=0.13
pyinstaller>=6.0
